Add files via upload
Download the 4 very small csv files first and save them in the same folder as the R code.  Then, run the R code to see optimized defensive positioning for Christian Yelich and Edwin Encarnacion.  There are also comments in the R script so you can follow along with each step.  This is a binary integer program using the lpSolve package in R.